+++ src/pages/play.page.route.ts
@@ -1,0 +1,4 @@
+// /pages/film.page.route.js
+
+// Match URLs `/film/1`, `/film/2`, ...
+export default '/play/:fileId'
