+++ src/pages/play.page.server.ts
@@ -1,0 +1,7 @@
+export { passToClient }
+
+// Example of `pageContext` often passed to the browser
+const passToClient = [
+  'pageProps',
+  'routeParams',
+]
